move {SLP, NAG} to their subdirectory
