+++ M2/Macaulay2/e/Makefile.files.in
@@ -101,8 +101,8 @@ INTERFACE = \
 	rings/tower \
 	interreduce \
 	finalize \
-	NAG \
-	SLP \
+	NAG/NAG \
+	SLP/SLP \
 	monomial/overflow \
 	memory-status \
 	imonorder \
