+++ M2/Macaulay2/e/NAG/NAG.cpp
