+++ M2/Macaulay2/e/SLP/SLP.cpp
